tests/requirements.txt
pandas~=2.3.3
PyPDF2~=3.0.1
allure-pytest